back/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.7.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
mysqlclient==2.1.1
PyJWT==2.4.0
pytz==2021.1
sqlparse==0.4.1

